+++ src/config/routes.js
@@ -2,14 +2,10 @@
 
 const controller = require('./config.controller');
 const openapi = require('../middleware/openapi');
+const orgAdmin = require('../middleware/identity/orgAdmin');
 
 module.exports = function (router) {
     router.get('/config', openapi('getConfig'), controller.get);
-    router.patch('/config', controller.requireOrgAdmin, openapi('updateConfig'), controller.patch);
-    router.delete(
-        '/config/:field',
-        controller.requireOrgAdmin,
-        openapi('deleteConfigField'),
-        controller.deleteConfig
-    );
+    router.patch('/config', orgAdmin, openapi('updateConfig'), controller.patch);
+    router.delete('/config/:field', orgAdmin, openapi('deleteConfigField'), controller.deleteConfig);
 };
